+++ pybamm/expression_tree/unary_operators.py
@@ -63,17 +63,17 @@ class UnaryOperator(pybamm.Symbol):
         """Perform unary operation on a child. """
         raise NotImplementedError
 
-    def evaluate(self, t=None, y=None, y_dot=None, params=None, known_evals=None):
+    def evaluate(self, t=None, y=None, y_dot=None, inputs=None, known_evals=None):
         """ See :meth:`pybamm.Symbol.evaluate()`. """
         if known_evals is not None:
             if self.id not in known_evals:
                 child, known_evals = self.child.evaluate(
-                    t, y, y_dot, params, known_evals
+                    t, y, y_dot, inputs, known_evals
                 )
                 known_evals[self.id] = self._unary_evaluate(child)
             return known_evals[self.id], known_evals
         else:
-            child = self.child.evaluate(t, y, y_dot, params)
+            child = self.child.evaluate(t, y, y_dot, inputs)
             return self._unary_evaluate(child)
 
     def _evaluate_for_shape(self):